# packages/userManagement/swagger.yaml
openapi: 3.0.3
info:
  title: iot-callbot-node
  description: |-

  contact:
    email: nguyenduyphuong3103@gmail.com
  version: 1.0.11
servers:
  - url: http://localhost:8000/api/
...
paths:
...
  /service/activateService:
    patch:
      tags:
        - service
      summary: Active service
      description: Active service
      operationId: activateService
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                serviceId:
                  type: string
                  example: "51087e3b-5d99-4e64-a2ec-3c474a5461fd"
                projectId:
                  type: string
                  example: "bd4855fb-0843-49eb-9427-8725bffc2538"
        required: true
      responses:
        default:
          description: found services
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    type: object
                    properties:
                      success:
                        type: boolean
                        example: true
                      message:
                        type: string
                        example: "Service has been activated !!!"
  /service/deactivateService:
    patch:
      tags:
        - service
      summary: Deactivate service
      description: Deactivate service
      operationId: deactivateService
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                serviceId:
                  type: string
                  example: "51087e3b-5d99-4e64-a2ec-3c474a5461fd"
                projectId:
                  type: string
                  example: "bd4855fb-0843-49eb-9427-8725bffc2538"
        required: true
      responses:
        default:
          description: found services
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    type: object
                    properties:
                      success:
                        type: boolean
                        example: true
                      message:
                        type: string
                        example: "Service has been deactivated !!!"
  /service/createServiceByUser/{projectId}/{serviceId}:
    post:
      tags:
        - service
      summary: Create service by userId
      description: "Create service by user"
      operationId: createServiceByUser
      security:
        - BearerAuth: []
      parameters:
        - name: projectId
          in: path
          description: "id of project"
          required: true
          schema:
            type: string
        - name: serviceId
          in: path
          description: "id of service"
          required: true
          schema:
            type: string
      responses:
        default:
          description: Created service by user
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    type: object
                    properties:
                      ok:
                        type: boolean
                        example: "true"
                      message:
                        type: string
                        example: "Service registration successfully!!!"
                  resData:
                    $ref: "#/components/schemas/ServiceByUser"
  /service/editServiceByUser/{projectId}/{serviceId}:
    patch:
      tags:
        - service
      summary: Edit service by userId
      description: "Edit service by user"
      operationId: editServiceByUser
      security:
        - BearerAuth: []
      parameters:
        - name: projectId
          in: path
          description: "id of project"
          required: true
          schema:
            type: string
        - name: serviceId
          in: path
          description: "id of service"
          required: true
          schema:
            type: string
      responses:
        default:
          description: Edited service by user
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    type: object
                    properties:
                      ok:
                        type: boolean
                        example: "true"
                      message:
                        type: string
                        example: "The data has been updated!!!"
...
components:
  schemas:
...
    ServiceByUser:
      type: object
      properties:
        name:
          type: string
          example: "service 1"
        sumData:
          type: number
          example: 0
        sumCost:
          type: number
          example: 0
        unpaid:
          type: number
          example: null
        inProject:
          type: object
          properties:
            id:
              type: string
              example: "bd4855fb-0843-49eb-9427-8725bffc2538"
      xml:
        name: service
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT